Enable default object tracking no Mac
We had trouble with base_unittests before, so we
ifdef'ed out the larger landing.  This tiny
switch CL will turn it on for the Mac, but can
be quickly reverted if there is a problem.

Scott Hess was able to debug and identify the
cause of the problem (yeah Scott!!!!).  I think
the code we landed corrects for the issue he identified,
and this should prove it (or get reverted trying).

r=rtenneti,shess
bug=101856
Review URL: http://codereview.chromium.org/8387059

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31b04d71f05924ee6ef914be7b7802bef73d5bf6