Avoid any possibility of an Alloc during TLS thread teardown
GIven that TCMalloc might get its thread teardown
notifgication ahead of the profiler, we need to be
careful to not do an alloc in the teardown handler.
If we did, we might bring TCMalloc back to life on
this thread... and it might not get a second teardown
notification.

r=rtenneti
BUG=104696
Review URL: http://codereview.chromium.org/8587031

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 445029fbd71d6c7b3fe145c26a76cbdc4c6613d3