Revert 113748 - Provide windows notifictaion of thread termination
[Karen Grunberg requested a speculative revert, because
we are currently unable to complete a link of the canary.
This change is working fine on the Dev branch, and did
not impact the link over the weekend, but for lack of
other items, this change is being reverted]


This code should only have much of an impact on the XP
platform, where thread teardown notification are not
currently made.

Provide an automatic fallback scan of the linker
list of notifcation callbacks via DllMain.

This fallback is ONLY used if there are no notifications
coming via the linker list. This allows all existing
code to continue working as it does today, and provides
thread teardown notifcatino on XP even though we use
the LoadLibrary call.

It is a minimal change, suitable for pushing to other
channels (beta? stable?).

r=cpu
BUG=103209
Review URL: http://codereview.chromium.org/8836003

TBR=jar@chromium.org
Review URL: http://codereview.chromium.org/8934006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a32da25057698ef8e113e2c9432e2b88164ab4ae