Redo r113722 - Add Pass(), which implements move semantics, to scoped_ptr, scoped_array....
-- This time for sure. --

Add Pass(), which implements move semantics, to scoped_ptr, scoped_array, and
scoped_ptr_malloc.

This modification to the scopers implements the "moveable but not copyable"
semantics that were introduced in C++11's unique_ptr<>.

With this, is now possible to use scopers as an argument type or a return type.
This signifies, in the type system, transfer of ownership into a function or out
of a function respectively.  Calling, or returning such a function MUST use the
temporary resulting from a function or explicit cast.

This distinction makes it possible to avoid the implicit ownership transfer
issues of auto_ptr, but still allow us to have compiler enforced ownership
transfer.

Also adds a Passed() helper that allows using a scoper with Bind().

Original Review URL: http://codereview.chromium.org/8774032

BUG=96118
TEST=new unittests

Review URL: http://codereview.chromium.org/9018037

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 206a2ae8a1ebd2b040753fff7da61bbca117757f