Add a new option to JSONWriter, which instructs it not to append '.0' or use exponential notation when writing doubles with no fractional part.
This is needed so we can write an integer value larger than what a 32-bit
int will hold using a double, and write the said double in such a way as
to enable external JSON readers (that support int64) to convert it to an
int64 instead of a double.

The other alternative is to add support for int64 in the Value classes and
clients. See http://codereview.chromium.org/8962042
BUG=none
TEST=none


Review URL: http://codereview.chromium.org/9016024

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ba4fb5addfd5c0c824491e8ae11743c5fc31cdb3