Added non-NULL DCHECK to MessageLoopForUI::current.
Having spent a little while tracking down a crash caused by not having a
message loop when calling this function, it seems that this is worth having.
I'm in two minds as to whether I prefer this approach to one where NULL is
a valid return value in debug builds as it is in release builds.

BUG=
TEST=


Review URL: http://codereview.chromium.org/9705073

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9600bd862ab0ce57d132f5cb039d702e11968cf2