mmap returns -1 on failure and in the posix implementation of SharedMemory, this was setting memory_ to -1. However the interface for SharedMemory specifies that memory_ should be NULL if it is unmapped.
Furthermore there was  DCHECK included to check the success of mmap, but I feel this is incorrect because we are returning the success (e.g. other platform implementations don't do this). The caller can DCHECK if they need. We could change this to a LOG call.

BUG=None
TEST=out/Debug/base_unittests


Review URL: http://codereview.chromium.org/10174007

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9ff39419c4f7be1a36a721a032146fe342bd54c4