Activate Keychain reauthorization.
If the KeychainReauthorizeInAppMay2012 preference doesn't exist or is less
than 2, Keychain reauthorization is performed at start-up. On successful
reauthorization, this preference will be set to 2. If unsuccessful, it will be
incremented by 1, giving a maximum of 2 unsuccessful attempts to perform this
procedure before getting on with life.

BUG=108238
TEST=none
Review URL: https://chromiumcodereview.appspot.com/10310174

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0c58de58e969865a768f042c0798a8a5ab93d8f