[Mac] Log stack trace for CHECK in bug 97285.
The hypothesis in comment #55 on the bug is that while setting up a
new channel to an already-connected plugin,
NPChannelBase::RemoveRoute() can be called on the earlier channel, so
the renderer and plugin channel mappings get out of sync.  This will
add a Breakpad key encoding the top stack frames when RemoveRoute() is
called within the scoped of WebPluginDelegateProxy::Initialize().

objc_zombies.mm already implemented the stackframe-encoding code, so I
pulled that to a central location.  That code can be left in place
when the other code is removed after some traces are generated.

BUG=97285
TEST=crash server.


Review URL: https://chromiumcodereview.appspot.com/10408004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c6aa6263aee292c2bd5d98f1cff1c5aceaec4f84