Base: Add power requirements to the System monitor.
The only thing we need in the short term is CPU_REQUIRED,
and that part is implemented for all platforms (as far as it
makes sense). The other options are implemented for Windows,
but given that currently there's no user, I expect the
implementation for other platforms to be written when there's
an actual need for it.

BUG=126591
TEST=none

Original review: https://chromiumcodereview.appspot.com/10406012/
TBR=brettw
Review URL: https://chromiumcodereview.appspot.com/10389227

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b9256cd092d5ebeae70806ef8b0f332dcea3b1d3