Cleanup: Remove the C version of GetUrandomFD()
I originally added this in a #includable-from-C header so that it
could be #included from NaCl-side code.  However, having NaCl #include
Chromium headers is not a clean interface, and this has since been
replaced by runtime dependency injection.

We move the function to a C++ header and put it into a namespace.

Also fix the error return value for UrandomFD() so that it can't be
confused with stdin.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2655
TEST=build

Review URL: https://chromiumcodereview.appspot.com/10446043

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ead8c1fa606c9a8712c8bb71bdb656f5357c717f