Fix a bug in file_util::Delete() where symlinks are not handled right
With this patch, file_util::Delete() uses lstat64() instead of stat64(), so that Delete() checks the existence of the symbolic link itself not a file pointed by the symlink. This allows us to delete a symbolic link when it points to a non-existent file. 

BUG=119430
TEST=base_unittests::FileUtilTest.* passes.

Review URL: https://chromiumcodereview.appspot.com/10690047

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7122db29d3965050505f7294f073d83f80ea7835