Exclude parts of the 'base' target's file list to get it building on iOS
This is a quick-and-dirty removal of everything that doesn't build as-is on iOS, so that we can get a bot running on the FYI waterfall as soon as possible. Once that's in place we'll unfork various files in base and remove the exclusions here.

(Note that we aren't building and running unit tests yet either; that will also come soon as a follow-up. Some of the code that's being built here still needs tweaking to actually work correctly, but since the first stage is just building something I only excluded things that failed to compile.)

BUG=None
TEST=None


Review URL: https://chromiumcodereview.appspot.com/10695039

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 304a6bfc33fb111c9109e589e579a883df2f47e3