[Chromoting] Launch the host process elevated via ShellExecuteEx().
The host process is launched in in two steps now. An instance of remote_service.exe is launched in a user session (CreateProcessAsUser) and then it launches the host requesting elevation (ShellExecuteEx). This is needed because Windows 8 refuses to inject Alt+Tab unless uiAccess='true' is specified in the manifest, which in its turn requires ShellExecuteEx() to be used.

Lifetime of launched processes is controlled by assigning them to a job object.

Message loop changes are required to be able to process job object notifications on the I/O message loop.

BUG=135217

Review URL: https://chromiumcodereview.appspot.com/10831271

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 773632872a2738c85cb4e29fdd300de02316fe26