linux: Crash browser on too-big messages to zygote.
This adds CHECKs to the browser if it attempts to send a
message to the zygote that exceeds the maximum message size
(which causes an EMSGSIZE error in the zygote) or that
contains too many file descriptors.

I'm hoping that this will help make the source of the
problem more apparent when we hit the message size limit,
which we appear to have done multiple times (it was
originally 1 KB and is now 2 KB).

BUG=154409


Review URL: https://chromiumcodereview.appspot.com/11108019

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 375a8463858537aa295b904a71073784dfa4de6a