Linux: inform the Zygote when it's waiting on a dead process
If the browser calls ProcessDied() and asks the Zygote to wait (without blocking)
on a dead process, the kernel might not be done destroying it and the Zygote may
mistakenly claim that the process is alive.

We now inform the Zygote over the IPC that the process is already dead so
that it can wait synchroneously.

BUG=157458
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/11316261

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7691de3400bc8d62d85c96d41323eee85b85d22