Assert on actual contents of ScopedVector in scoped_vector_unittest.
It's great if objects have the expected life cycles, and it's great if the vectors have the right sizes and clear at the right times.

But isn't it also nice if the expected object is being stored in the expected place?

R=willchan@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/11316156

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e31354232246bf18820115ea1d4cc4d341a09ce6