Add a base::ParseJson() function to help tests construct Values.
This is very similar to base::JSONReader::Read() except that it allows
trailing commas and EXPECT-fails with a good message when the json is
malformed.

I migrated two tests to use this as a demonstration that it's widely
useful and makes things more readable.


Review URL: https://chromiumcodereview.appspot.com/11567027

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 06d19d8a47db854b256427066b70c91fffa1775e