Fix flakyness in PrefMemberTest::MoveToThread
 In MoveToThread, a thread object is owned by a ref-counted object that
is passed around thread. It means the thread object can be deleted on
a thread that is not the one that created it, which will trigger a
DCHECK. This CL ensures that the second thread is always stopped on the main
thread.

R=battre@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/11624023

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8dda17e4af70393f93d8f29c1e574907ebac53b8